Водительская часть Выделена в отдельный модуль
