+++ src/main/java/com/shippings/controllers/TestController.java
@@ -32,50 +32,4 @@ public class TestController {
         return "Client Content.";
     }
 
-    @Autowired
-    private ShippRepository shippRepository;
-
-    private static final Logger LOG = LoggerFactory.getLogger(UserController.class);
-
-    long getCurrentUserId() {
-        UserDetailsImpl user = (UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal();;
-        return user.getId();
-    }
-
-    @GetMapping("/driver")
-    @PreAuthorize("hasAuthority('driver')")
-    public List<Map<String, String>> driverAccess(){
-        LOG.info("driver request is received");
-        return getRouteList(Boolean.TRUE);
-    }
-
-    @GetMapping("/driverRequest")
-    @PreAuthorize("hasAuthority('driver')")
-    public List<Map<String, String>> driverRequestAccess() {
-        LOG.info("driver request request is received.");
-       return getRouteList(Boolean.FALSE);
-    }
-
-    List<Map<String, String>> getRouteList(Boolean bool){
-        List<Map<String, String>> shippList = new ArrayList();
-        List<Shipping> shipps = shippRepository.findAllByDriverIdAndStatus(this.getCurrentUserId(), bool);
-        LOG.info(String.format("shipp.size is %d", shipps.size()));
-        for (Shipping s : shipps)
-            shippList.add(new HashMap<String, String>(){{
-                put("id", Long.toString(s.getId()));
-                put("date_start", s.getDateStart().toString());
-                put("date_finish", s.getDateFinish().toString());
-                put("start", s.getStart());
-                put("finish", s.getFinish());
-                put("weight", Integer.toString(s.getWeight()));
-                put("height", Integer.toString(s.getHeight()));
-                put("length", Integer.toString(s.getLength()));
-                put("width", Integer.toString(s.getWidth()));
-                put("plus_time", Integer.toString(s.getPlusTime()));
-                put("comment", s.getComment());
-            }});
-
-        return shippList;
-    }
-
 }
