+++ src/main/java/com/shippings/repositories/ShippRepository.java
@@ -8,6 +8,7 @@ import java.util.List;
 
 @Repository
 public interface ShippRepository extends JpaRepository<Shipping, Long> {
-    List<Shipping> findAllByDriverIdAndStatus(Long driverId, Boolean bool);
+    List<Shipping> findAllByDriverIdAndStatusOrderByDateStartDesc(Long driverId, Boolean bool);
+    Shipping findOneById(Long ID);
 
 }
